requirements.txt
einops
fvcore
omegaconf
hydra-core
scipy
h5py
python-dateutil
lightning
tqdm
pandas
mne
nvidia-nccl-cu11
psutil
pyyaml
tensorboardX
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
scikit-learn
tensorboard
tensorrt
torch_optimizer
tables
timm
torchaudio
torcheeg
torchmetrics
warmup-scheduler